+++ server/routes.js
@@ -43,33 +43,5 @@ exports.getApp = (req, res) => {
 }
 
 
-// app.get ('/home',function (req, res) =>{
-//   db.getStyles(data => {
-//     res.status(200).send(data)
-//   })
-// });
-//
-// app.get  ('/style', function (req,res) => {
-//   Promise.all([
-//     db.getInfoByStyle(),
-//     db.imagesByStyle()
-//   ])
-//     .then (data => {
-//       res.status(200).send(data)
-//     })
-//     .catch (err => {
-//       console.log err;
-//     })
-// })
-//
-// app.post ('/login',function (req, res) => {
-//   db.getUser(data => {
-//     res.status(200).send(data)
-//   })
-// });
-//
-// app.post ('/login',function (req, res) => {
-//   db.getApp(data => {
-//     res.status(200).send(data)
-//   })
-// });
+
+
